+++ docs/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+## [8.1.0] - 2024-07-15
+
 ### Fixed
 
 -   Fixed a bug with some commands being calculated twice
@@ -462,7 +464,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
-[Unreleased]: https://github.com/spapanik/yamk/compare/v8.0.0...main
+[Unreleased]: https://github.com/spapanik/yamk/compare/v8.1.0...main
+[8.1.0]: https://github.com/spapanik/yamk/compare/v8.0.0...v8.1.0
 [8.0.0]: https://github.com/spapanik/yamk/compare/v7.0.0...v8.0.0
 [7.0.0]: https://github.com/spapanik/yamk/compare/v6.1.0...v7.0.0
 [6.1.0]: https://github.com/spapanik/yamk/compare/v6.0.1...v6.1.0